+++ django_node/npm.py
@@ -52,7 +52,7 @@ def run(*args):
     )
 
 
-def install(target_dir, arguments=None, silent=None):
+def install(target_dir, *args, **kwargs):
     """
     A method that will invoke `npm install` in a specified directory. Optional arguments will be
     appended to the invoked command.
@@ -60,8 +60,8 @@ def install(target_dir, arguments=None, silent=None):
     Arguments:
 
     - `target_dir`: a string pointing to the directory which the command will be invoked in.
-    - `arguments`: an optional tuple of strings to append to the invoked command.
-    - `silent`: an optional boolean indicating that NPM's output should not be printed to the terminal.
+    - `*args`: optional strings to append to the invoked command.
+    - `silent`: an optional keyword argument indicating that NPM's output should not be printed to the terminal.
 
     ```
     from django_node import npm
@@ -70,7 +70,10 @@ def install(target_dir, arguments=None, silent=None):
     stderr, stdout = npm.install('/path/to/some/directory/')
 
     # Install a dependency into a particular directory and persist it to the package.json file
-    stderr, stdout = npm.install('/path/to/some/directory/', ('--save', 'some-package'))
+    stderr, stdout = npm.install('/path/to/some/directory/', '--save', 'some-package')
+
+    # Install dependencies but suppress NPM's output
+    stderr, stdout = npm.install('/path/to/some/directory/', silent=False)
     ```
     """
 
@@ -79,12 +82,16 @@ def install(target_dir, arguments=None, silent=None):
             'npm.install\'s `target_dir` parameter must be either a str or unicode. Received: {0}'.format(target_dir)
         )
 
-    if arguments is None:
-        arguments = ()
-    elif not isinstance(arguments, tuple):
+    if not os.path.exists(target_dir) or not os.path.isdir(target_dir):
         raise exceptions.NpmInstallArgumentsError(
-            'npm.install\'s `arguments` parameter must be either a tuple or None. Received: {0}'.format(arguments)
+            'npm.install\'s `target_dir` parameter must be a string pointing to a directory. Received: {0}'.format(
+                target_dir
+            )
         )
+
+    args = tuple(args)
+
+    silent = kwargs.get('silent', False)
 
     ensure_installed()
 
@@ -92,7 +99,7 @@ def install(target_dir, arguments=None, silent=None):
 
     os.chdir(target_dir)
 
-    stderr, stdout = run(settings.NPM_INSTALL_COMMAND, *arguments)
+    stderr, stdout = run(settings.NPM_INSTALL_COMMAND, *args)
 
     if stderr:
         for line in stderr.splitlines():
@@ -106,7 +113,7 @@ def install(target_dir, arguments=None, silent=None):
         print('-' * 80)
         print(
             'Output from `{command}` in {target_dir}'.format(
-                command=' '.join((settings.PATH_TO_NPM,) + arguments),
+                command=' '.join((settings.PATH_TO_NPM,) + args),
                 target_dir=target_dir,
             )
         )